+++ packages/react-components/react-components/stories/ContentEditableTags/Lexical2/AutocompletePlugin.tsx
@@ -1,24 +1,10 @@
 import { makeStyles, mergeClasses, shorthands, useAnnounce } from '@fluentui/react-components';
-import { useLexicalComposerContext } from '@lexical/react/LexicalComposerContext';
-import { people } from '../data';
-import {
-  $getSelection,
-  SELECTION_CHANGE_COMMAND,
-  KEY_ENTER_COMMAND,
-  COMMAND_PRIORITY_EDITOR,
-  $isRangeSelection,
-  COMMAND_PRIORITY_CRITICAL,
-  INSERT_PARAGRAPH_COMMAND,
-  KEY_ARROW_UP_COMMAND,
-  KEY_ARROW_DOWN_COMMAND,
-  $isNodeSelection,
-} from 'lexical';
 import * as React from 'react';
-import { $createNamePillNode } from './NamePillNode';
+import { people } from '../data';
+import { AutocompletePluginCore } from './AutocompletePluginCore';
 
 const useStyles = makeStyles({
   root: {
-    // Stack the label above the field with a gap
     display: 'grid',
     gridTemplateRows: 'repeat(1fr)',
     justifyItems: 'start',
@@ -42,101 +28,6 @@ const useFilteredList = (filter: string) => {
   return options.filter(option => option.toLowerCase().includes(filter.toLowerCase()));
 };
 
-export type AutocompletePluginProps = {
-  query: string;
-};
-
-type AutocompletePluginLexicalHooksProps = {
-  onQueryChange: (newQuery: string) => void;
-  newPillCandidate?: string;
-  onKeyUp: (event: KeyboardEvent) => boolean;
-  onKeyDown: (event: KeyboardEvent) => boolean;
-  children: (renderProps: { appendCandidate: () => void }) => React.ReactElement;
-};
-
-const AutocompletePluginLexicalHooks: React.FC<AutocompletePluginLexicalHooksProps> = ({
-  onQueryChange,
-  newPillCandidate,
-  children,
-  onKeyUp,
-  onKeyDown,
-}) => {
-  const [editor] = useLexicalComposerContext();
-
-  const appendSelectedItem = React.useCallback(() => {
-    const sel = $getSelection();
-
-    // handle deletion of single selected node
-    if ($isNodeSelection(sel) && sel.getNodes().length === 1) {
-      const selectedNode = sel.getNodes()[0];
-      selectedNode.remove();
-      return true;
-    }
-
-    // handle adding of new node from selection
-    if ($isRangeSelection(sel) && sel.getNodes().length === 1) {
-      if (newPillCandidate) {
-        const node = sel.getNodes()[0];
-        const newNode = $createNamePillNode(newPillCandidate);
-        node.replace(newNode);
-        newNode.selectEnd();
-        return true;
-      }
-    }
-    return false;
-  }, [editor, newPillCandidate]);
-
-  React.useEffect(() => {
-    return editor.registerCommand(KEY_ENTER_COMMAND, appendSelectedItem, COMMAND_PRIORITY_CRITICAL);
-  });
-
-  React.useEffect(() => {
-    return editor.registerCommand(
-      SELECTION_CHANGE_COMMAND,
-      () => {
-        const sel = $getSelection();
-        if ($isRangeSelection(sel) && sel.getNodes().length === 1) {
-          const node = sel.getNodes()[0];
-          onQueryChange(node.getTextContent());
-        }
-      },
-      COMMAND_PRIORITY_EDITOR,
-    );
-  }, [editor]);
-
-  React.useEffect(() => {
-    return editor.registerCommand(
-      INSERT_PARAGRAPH_COMMAND,
-      () => {
-        return true;
-      },
-      COMMAND_PRIORITY_CRITICAL,
-    );
-  }, [editor]);
-
-  React.useEffect(() => {
-    return editor.registerCommand(
-      KEY_ARROW_UP_COMMAND,
-      payload => {
-        return onKeyUp(payload);
-      },
-      COMMAND_PRIORITY_CRITICAL,
-    );
-  }, [editor, onKeyUp]);
-
-  React.useEffect(() => {
-    return editor.registerCommand(
-      KEY_ARROW_DOWN_COMMAND,
-      payload => {
-        return onKeyDown(payload);
-      },
-      COMMAND_PRIORITY_CRITICAL,
-    );
-  }, [editor, onKeyDown]);
-
-  return children({ appendCandidate: appendSelectedItem });
-};
-
 export const AutocompletePlugin = () => {
   const [isOpen, setIsOpen] = React.useState(true);
   const [query, setQuery] = React.useState('');
@@ -147,13 +38,8 @@ export const AutocompletePlugin = () => {
   const filtered = useFilteredList(query);
 
   React.useEffect(() => {
-    if (query.length) {
-      setIsOpen(true);
-      setSelectedIndex(0);
-    } else {
-      setIsOpen(false);
-      setSelectedIndex(0);
-    }
+    setSelectedIndex(0);
+    setIsOpen(!!query.length);
   }, [query]);
 
   React.useEffect(() => {
@@ -204,7 +90,7 @@ export const AutocompletePlugin = () => {
   );
 
   return (
-    <AutocompletePluginLexicalHooks
+    <AutocompletePluginCore
       onQueryChange={newQuery => setQuery(newQuery)}
       newPillCandidate={filtered[selectedIndex]}
       onKeyUp={onKeyUp}
@@ -228,6 +114,6 @@ export const AutocompletePlugin = () => {
           </div>
         );
       }}
-    </AutocompletePluginLexicalHooks>
+    </AutocompletePluginCore>
   );
 };
